LayoutTests/canvas/philip/tests/2d.state.saverestore.stack.html

<!DOCTYPE html>
<title>Canvas test: 2d.state.saverestore.stack</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.lineWidth = 1;
ctx.save();
ctx.lineWidth = 2;
ctx.save();
ctx.lineWidth = 3;
_assertEqual(ctx.lineWidth, 3, "ctx.lineWidth", "3");
ctx.restore();
_assertEqual(ctx.lineWidth, 2, "ctx.lineWidth", "2");
ctx.restore();
_assertEqual(ctx.lineWidth, 1, "ctx.lineWidth", "1");


});
</script>
